Added DTO and lower case id for addressbook

+++ src/main/java/com/naimish/AddressBook/repository/AddressBookRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AddressBookRepository extends JpaRepository<AddressBook, Integer> {
+public interface AddressBookRepository extends JpaRepository<AddressBook, Long> {
 }
